refactor(web): app.js から provider-usage / terminal-attachments / index-search surface を切り出す
SPEC-3064 Phase 3 (E1-E3, arch-review C1)。app.js 16,120 → 14,232 行:

- provider-usage-surface.js (477): 使用量フォーマッタ・カード・hover
- terminal-attachments.js (852): clipboard / paste / drop / 添付進捗 +
  installer 群（terminal-copy-shortcut / context-menu の import も移管）
- project-index-search-surface.js (744): index 検索 state・レンダ・
  Index window mount

deps-bag factory（createWorkspaceOverviewSurface 前例）で配線し、
app.js 側は同名 const 受けで呼び出しテキスト不変。root_js_modules! /
Playwright ROOT_MODULES / check-frontend-bundle.sh へ 3 ファイル登録。
ソース固定テスト（index-status-controller 等）は読み先を新モジュールへ
付け替え（意図不変）。

検証: run-frontend-unit-tests 797/797、check-frontend-bundle PASS、
cargo test -p gwt --bin gwt embedded_web 104/104。
