+++ crates/gwt/web/app.js
@@ -27,14 +27,15 @@
         visibleBoardEntries,
       } from "/board-surface.js";
       import { createWorkspaceKanbanSurface as createWorkspaceOverviewSurface } from "/workspace-kanban-surface.js";
+      import { createProviderUsageSurface } from "/provider-usage-surface.js";
+      import { createTerminalAttachments } from "/terminal-attachments.js";
+      import { createProjectIndexSearchSurface } from "/project-index-search-surface.js";
       import { createWorkspaceResumePickerController } from "/workspace-resume-picker-modal.js";
       import { createLaunchPendingController } from "/launch-pending-controller.js";
       import { createConnectionOverlay } from "/connection-overlay.js";
       import { createUpdateCtaController } from "/update-cta.js";
       import { createReleaseNotesWindow } from "/release-notes-window.js";
       import { createConsoleWindow } from "/console-window.js";
-      import { createTerminalContextMenuController } from "/terminal-context-menu.js";
-      import { classifyTerminalCopyKeyEvent } from "/terminal-copy-shortcut.js";
       import { createTerminalWheelScrollController } from "/terminal-wheel-scroll.js";
       import {
         renderProjectTabs as renderProjectTabsView,
@@ -232,8 +233,6 @@
       const boardStateMap = new Map();
       const logStateMap = new Map();
       const knowledgeBridgeStateMap = new Map();
-      const indexSearchStateMap = new Map();
-      const pendingIndexOpenTargetsByPreset = new Map();
       const KNOWLEDGE_AUTO_REFRESH_INTERVAL_MS = 60000;
       let nextKnowledgeLoadRequestId = 1;
       let nextKnowledgeSearchRequestId = 1;
@@ -900,31 +899,37 @@
       };
       let versionState = { current: "", latest: "" };
       let pendingAboutHashOpen = false;
-      const indexStatusByProjectRoot = new Map();
       let projectError = "";
       const TERMINAL_SELECTION_DRAG_THRESHOLD = 4;
 
-      // SPEC-1939 Phase 13: project-bar Index badge withdrawn. The
-      // aggregated payload now feeds only the per-tab dot indicator and the
-      // Settings.Index panel; the project-bar surface no longer renders an
-      // Index summary because repo-shared (issues/specs) and per-worktree
-      // (files/files-docs) scopes were being collapsed into a single state.
-      function setIndexStatus(projectRoot, status) {
-        if (!projectRoot) {
-          return;
-        }
-        indexStatusByProjectRoot.set(projectRoot, {
-          state: status?.state || "",
-          detail: status?.detail || "",
-          repair_started_at: status?.repair_started_at || null,
-          progress: status?.progress || null,
-          scopes: status?.scopes || {},
-          worktrees: status?.worktrees || {},
-        });
-        renderIndexPanelInAllSettingsWindows();
-        renderProjectIndexWindows();
-        refreshProjectTabDots();
-      }
+      // SPEC-3064 Phase 3 (E3): the Project Index window surface (per-project
+      // index status map, Index search state, render/search/open helpers, and
+      // the Index window mount) moved to /project-index-search-surface.js.
+      // app.js keeps the receive(), Settings, knowledge-mount, and
+      // window-cleanup call sites and wires them through this factory.
+      const {
+        setIndexStatus,
+        handleProjectIndexSearchResults,
+        handleProjectIndexSearchError,
+        mountProjectIndexSurface,
+        indexSearchStateMap,
+        pendingIndexOpenTargetsByPreset,
+        indexStatusByProjectRoot,
+      } = createProjectIndexSearchSurface({
+        send,
+        sendWindowFocus: (id) => socketTransport.send({ kind: "focus_window", id }),
+        focusWindowLocally,
+        activeProjectTab,
+        makeEl,
+        clearChildren,
+        focusOrSpawnPreset,
+        knowledgeKindForPreset,
+        requestKnowledgeDetail,
+        renderKnowledgeBridge,
+        renderIndexPanelInAllSettingsWindows,
+        refreshProjectTabDots,
+        requestFullIndexStatusRefresh,
+      });
 
       function formatVersionLabel() {
         const current = versionState.current;
@@ -1009,518 +1014,6 @@
           return preset;
         }
         return null;
-      }
-
-      const INDEX_SEARCH_SCOPES = Object.freeze([
-        { id: "issues", label: "Issues" },
-        { id: "specs", label: "SPECs" },
-        { id: "board", label: "Board" },
-        { id: "discussions", label: "Discussions" },
-        { id: "files", label: "Files" },
-        { id: "files-docs", label: "Docs" },
-        { id: "memory", label: "Memory" },
-      ]);
-      const INDEX_SEARCH_DEFAULT_SCOPES = Object.freeze([
-        "issues",
-        "specs",
-        "board",
-        "discussions",
-        "memory",
-      ]);
-      function ensureIndexSearchState(windowId) {
-        if (!indexSearchStateMap.has(windowId)) {
-          indexSearchStateMap.set(windowId, {
-            activeTab: "search",
-            query: "",
-            matchMode: "semantic",
-            selectedScopes: new Set(INDEX_SEARCH_DEFAULT_SCOPES),
-            selectedWorktreeHash: "",
-            searchTimer: 0,
-            requestId: 0,
-            inFlightRequestId: 0,
-            inFlightSignature: "",
-            searching: false,
-            results: [],
-            suggestions: [],
-            selectedResultIndex: -1,
-            error: "",
-          });
-        }
-        return indexSearchStateMap.get(windowId);
-      }
-
-      function invalidateProjectIndexSearchRequest(state) {
-        state.requestId += 1;
-        state.inFlightRequestId = 0;
-        state.inFlightSignature = "";
-      }
-
-      function clearProjectIndexSearchState(state) {
-        if (state.searchTimer) {
-          clearTimeout(state.searchTimer);
-          state.searchTimer = 0;
-        }
-        invalidateProjectIndexSearchRequest(state);
-        state.results = [];
-        state.suggestions = [];
-        state.selectedResultIndex = -1;
-        state.searching = false;
-        state.error = "";
-      }
-
-      function markProjectIndexSearchPending(state) {
-        invalidateProjectIndexSearchRequest(state);
-        state.error = "";
-      }
-
-      function activeIndexStatus() {
-        const activeProjectRoot = activeProjectTab()?.project_root || "";
-        return (activeProjectRoot && indexStatusByProjectRoot.get(activeProjectRoot)) || null;
-      }
-
-      function indexWorktreeEntries(status) {
-        const worktrees = status?.worktrees || {};
-        return Object.entries(worktrees)
-          .map(([hash, meta]) => ({
-            hash,
-            branch: meta?.branch || "",
-            path: meta?.path || "",
-            label: meta?.branch || meta?.path || hash,
-          }))
-          .sort((left, right) => left.label.localeCompare(right.label));
-      }
-
-      function activeWorktreeHashForIndex(status) {
-        const activeProjectRoot = activeProjectTab()?.project_root || "";
-        const entries = indexWorktreeEntries(status);
-        return (
-          entries.find((entry) => entry.path === activeProjectRoot)?.hash
-          || entries[0]?.hash
-          || ""
-        );
-      }
-
-      function selectedIndexWorktreeHash(state, status) {
-        const entries = indexWorktreeEntries(status);
-        if (state.selectedWorktreeHash && entries.some((entry) => entry.hash === state.selectedWorktreeHash)) {
-          return state.selectedWorktreeHash;
-        }
-        return activeWorktreeHashForIndex(status);
-      }
-
-      function indexFileScopesSelected(state) {
-        return state.selectedScopes.has("files") || state.selectedScopes.has("files-docs");
-      }
-
-      function selectedIndexScopeLabels(state) {
-        return INDEX_SEARCH_SCOPES
-          .filter((scope) => state.selectedScopes.has(scope.id))
-          .map((scope) => scope.label);
-      }
-
-      function formatIndexSearchMatch(distance) {
-        if (!Number.isFinite(distance)) return "";
-        const score = Math.max(0, Math.min(1, 1 - distance));
-        return `${Math.round(score * 100)}% match`;
-      }
-
-      function indexSearchVisibleResults(state) {
-        return [
-          ...state.results.map((result) => ({ result, suggestion: false })),
-          ...state.suggestions.map((result) => ({ result, suggestion: true })),
-        ];
-      }
-
-      function selectedIndexSearchItem(state) {
-        const visible = indexSearchVisibleResults(state);
-        return visible[state.selectedResultIndex] || visible[0] || null;
-      }
-
-      function formatIndexSearchEvidence(result, includeMissing = false) {
-        const matched = Array.isArray(result?.matched_terms) ? result.matched_terms : [];
-        const missing = Array.isArray(result?.missing_terms) ? result.missing_terms : [];
-        const parts = [];
-        if (matched.length > 0) {
-          parts.push(`Matched: ${matched.join(", ")}`);
-        }
-        if (includeMissing && missing.length > 0) {
-          parts.push(`Missing: ${missing.join(", ")}`);
-        }
-        return parts.join(" · ");
-      }
-
-      function indexSearchLoadingLabel(state) {
-        return state.matchMode === "all_terms" ? "Searching all terms" : "Searching semantic index";
-      }
-
-      function indexSearchPlaceholder(state) {
-        return state.matchMode === "all_terms"
-          ? "All terms required, e.g. Work discussion"
-          : "Search by meaning, e.g. work lifecycle";
-      }
-
-      function scheduleProjectIndexSearch(windowId) {
-        const state = ensureIndexSearchState(windowId);
-        if (state.searchTimer) {
-          clearTimeout(state.searchTimer);
-        }
-        state.searchTimer = setTimeout(() => {
-          state.searchTimer = 0;
-          sendProjectIndexSearch(windowId);
-        }, 250);
-      }
-
-      function sendProjectIndexSearch(windowId) {
-        const state = ensureIndexSearchState(windowId);
-        const query = state.query.trim();
-        if (!query) {
-          clearProjectIndexSearchState(state);
-          renderProjectIndexSearch(windowId);
-          return;
-        }
-        const scopes = Array.from(state.selectedScopes);
-        if (scopes.length === 0) {
-          invalidateProjectIndexSearchRequest(state);
-          state.error = "Select at least one scope.";
-          state.searching = false;
-          renderProjectIndexSearch(windowId);
-          return;
-        }
-        const requestId = state.requestId + 1;
-        const status = activeIndexStatus();
-        const worktreeHash = indexFileScopesSelected(state)
-          ? selectedIndexWorktreeHash(state, status)
-          : "";
-        const matchMode = state.matchMode || "semantic";
-        const searchSignature = JSON.stringify({ query, scopes, worktreeHash, matchMode });
-        if (state.searching && state.inFlightSignature === searchSignature) {
-          renderProjectIndexSearch(windowId);
-          return;
-        }
-        state.requestId = requestId;
-        state.inFlightRequestId = requestId;
-        state.inFlightSignature = searchSignature;
-        state.searching = true;
-        state.error = "";
-        send({
-          kind: "search_project_index",
-          id: windowId,
-          query,
-          request_id: requestId,
-          scopes,
-          match_mode: state.matchMode,
-          worktree_hash: worktreeHash || null,
-        });
-        renderProjectIndexSearch(windowId);
-      }
-
-      function renderProjectIndexSearch(windowId) {
-        const state = ensureIndexSearchState(windowId);
-        const root = document.querySelector(
-          `[data-id='${CSS.escape(windowId)}'] .index-search-root`,
-        );
-        if (!root) return;
-        const status = activeIndexStatus();
-        const searchPanel = root.querySelector("[data-index-panel='search']");
-        const healthPanel = root.querySelector("[data-index-panel='health']");
-        const healthTable = root.querySelector(".index-health-table");
-        const tabs = root.querySelectorAll("[data-index-tab]");
-        tabs.forEach((tab) => {
-          const active = tab.dataset.indexTab === state.activeTab;
-          tab.classList.toggle("active", active);
-          tab.setAttribute("aria-selected", String(active));
-        });
-        if (searchPanel) {
-          searchPanel.hidden = state.activeTab !== "search";
-        }
-        if (healthPanel) {
-          healthPanel.hidden = state.activeTab !== "health";
-        }
-        renderProjectIndexSearchControls(root, state, status);
-        renderProjectIndexSearchResults(root, state);
-        if (healthTable) {
-          renderIndexSettingsPanel({
-            panel: healthTable,
-            status,
-            projectRoot: activeProjectTab()?.project_root || "",
-            send,
-          });
-        }
-      }
-
-      function renderProjectIndexSearchControls(root, state, status) {
-        const scopes = root.querySelector(".index-scope-list");
-        if (scopes) {
-          clearChildren(scopes);
-          for (const scope of INDEX_SEARCH_SCOPES) {
-            const button = makeEl("button", {
-              className: "index-scope-button",
-              attrs: {
-                type: "button",
-                "aria-pressed": String(state.selectedScopes.has(scope.id)),
-              },
-              dataset: { scope: scope.id },
-              text: scope.label,
-            });
-            scopes.appendChild(button);
-          }
-        }
-        const matchModes = root.querySelectorAll("[data-match-mode]");
-        matchModes.forEach((button) => {
-          const active = button.dataset.matchMode === state.matchMode;
-          button.classList.toggle("active", active);
-          button.setAttribute("aria-pressed", String(active));
-        });
-        const runButton = root.querySelector(".index-run-button");
-        if (runButton) {
-          runButton.disabled = !state.query.trim() || state.searching;
-          runButton.textContent = state.searching ? "Searching" : "Search";
-        }
-        const input = root.querySelector(".index-search-input");
-        if (input) {
-          input.placeholder = indexSearchPlaceholder(state);
-        }
-        const fileWorktreeField = root.querySelector(".index-worktree-field");
-        if (fileWorktreeField) {
-          fileWorktreeField.hidden = !indexFileScopesSelected(state);
-        }
-        const select = root.querySelector(".index-worktree-select");
-        if (select) {
-          clearChildren(select);
-          select.disabled = !indexFileScopesSelected(state);
-          const entries = indexWorktreeEntries(status);
-          const selectedHash = selectedIndexWorktreeHash(state, status);
-          if (entries.length === 0) {
-            select.appendChild(makeEl("option", { attrs: { value: "" }, text: "Current workspace" }));
-          } else {
-            for (const entry of entries) {
-              select.appendChild(
-                makeEl("option", {
-                  attrs: { value: entry.hash },
-                  text: entry.label,
-                }),
-              );
-            }
-          }
-          select.value = selectedHash;
-        }
-        const statusNode = root.querySelector(".index-search-status");
-        if (statusNode) {
-          const scopeSummary = selectedIndexScopeLabels(state).join(", ");
-          if (state.searching) {
-            statusNode.textContent = state.results.length > 0
-              ? `Updating results · ${scopeSummary}`
-              : `${indexSearchLoadingLabel(state)} · ${scopeSummary}`;
-          } else if (state.error) {
-            statusNode.textContent = state.error;
-          } else if (state.query.trim() && indexSearchVisibleResults(state).length > 0) {
-            statusNode.textContent = state.matchMode === "all_terms"
-              ? `${state.results.length} strict results · ${state.suggestions.length} semantic suggestions · ${scopeSummary}`
-              : `${state.results.length} results · ${scopeSummary}`;
-          } else if (state.query.trim()) {
-            statusNode.textContent = "No indexed results";
-          } else {
-            statusNode.textContent = `Ready · ${scopeSummary}`;
-          }
-          statusNode.dataset.kind = state.error ? "error" : "";
-        }
-      }
-
-      function renderProjectIndexSearchResults(root, state) {
-        const layout = root.querySelector(".index-search-layout");
-        const list = root.querySelector(".index-result-list");
-        const detail = root.querySelector(".index-result-detail");
-        if (!list || !detail) return;
-        const visibleResults = indexSearchVisibleResults(state);
-        clearChildren(list);
-        clearChildren(detail);
-        layout?.classList.toggle("is-empty", visibleResults.length === 0);
-        if (layout) {
-          layout.setAttribute("aria-busy", String(Boolean(state.searching)));
-        }
-        if (!state.query.trim()) {
-          list.appendChild(makeEl("div", { className: "workspace-empty-state", text: "Search indexed content." }));
-        } else if (state.searching && state.results.length === 0) {
-          list.appendChild(makeIndexSearchLoadingState(indexSearchLoadingLabel(state)));
-        } else if (state.error) {
-          list.appendChild(makeEl("div", { className: "workspace-empty-state", text: state.error }));
-        } else if (visibleResults.length === 0) {
-          list.appendChild(makeEl("div", { className: "workspace-empty-state", text: "No indexed results" }));
-        } else {
-          if (state.searching) {
-            list.appendChild(makeIndexSearchLoadingState("Updating results"));
-          }
-          let rowIndex = 0;
-          const appendResultGroup = (label, items, suggestion) => {
-            if (items.length === 0) return;
-            if (state.matchMode === "all_terms") {
-              list.appendChild(makeEl("div", { className: "index-result-group-label", text: label }));
-            }
-            items.forEach((result) => {
-              const index = rowIndex;
-              rowIndex += 1;
-              const row = makeEl("button", {
-                className: `index-result-row${suggestion ? " is-suggestion" : ""}`,
-                attrs: {
-                  type: "button",
-                  "aria-selected": String(index === state.selectedResultIndex),
-                },
-                dataset: { resultIndex: String(index) },
-              });
-              row.appendChild(makeEl("span", { className: "index-result-scope", text: result.scope || "" }));
-              row.appendChild(makeEl("strong", { text: result.title || "Untitled" }));
-              row.appendChild(makeEl("span", {
-                className: "index-result-subtitle",
-                text: [result.subtitle || "", formatIndexSearchMatch(result.distance)].filter(Boolean).join(" · "),
-              }));
-              const evidence = formatIndexSearchEvidence(result);
-              if (evidence) {
-                row.appendChild(makeEl("span", { className: "index-result-evidence", text: evidence }));
-              }
-              list.appendChild(row);
-            });
-          };
-          appendResultGroup("Strict results", state.results, false);
-          appendResultGroup("Semantic suggestions", state.suggestions, true);
-        }
-
-        const selectedItem = selectedIndexSearchItem(state);
-        if (!selectedItem) {
-          detail.appendChild(makeEl("div", { className: "workspace-empty-state", text: "Select a result" }));
-          return;
-        }
-        const selected = selectedItem.result;
-        state.selectedResultIndex = Math.max(
-          0,
-          Math.min(visibleResults.length - 1, state.selectedResultIndex),
-        );
-        detail.appendChild(makeEl("h3", { className: "index-detail-title", text: selected.title || "Untitled" }));
-        detail.appendChild(makeEl("div", { className: "index-detail-subtitle", text: selected.subtitle || selected.scope || "" }));
-        const match = formatIndexSearchMatch(selected.distance);
-        if (match) {
-          detail.appendChild(makeEl("div", { className: "index-detail-meta", text: match }));
-        }
-        const evidence = formatIndexSearchEvidence(selected, true);
-        if (evidence) {
-          detail.appendChild(makeEl("div", { className: "index-detail-meta", text: evidence }));
-        }
-        detail.appendChild(makeEl("p", { className: "index-detail-preview", text: selected.preview || "No preview available" }));
-        detail.appendChild(makeEl("button", {
-          className: "wizard-button primary",
-          attrs: { type: "button" },
-          dataset: { action: "open-index-result" },
-          text: "Open",
-        }));
-      }
-
-      function makeIndexSearchLoadingState(label = "Searching semantic index") {
-        const node = makeEl("div", {
-          className: "index-search-loading",
-          attrs: { role: "status", "aria-live": "polite" },
-        });
-        const dots = makeEl("span", {
-          className: "index-search-loading-dots",
-          attrs: { "aria-hidden": "true" },
-        });
-        for (let i = 0; i < 3; i += 1) {
-          dots.appendChild(makeEl("span", { className: "index-search-loading-dot" }));
-        }
-        node.appendChild(dots);
-        node.appendChild(makeEl("span", { className: "index-search-loading-label", text: label }));
-        return node;
-      }
-
-      function handleProjectIndexSearchResults(event) {
-        const state = indexSearchStateMap.get(event.id);
-        if (!state || event.request_id !== state.inFlightRequestId) {
-          return;
-        }
-        state.searching = false;
-        state.inFlightRequestId = 0;
-        state.inFlightSignature = "";
-        state.error = "";
-        state.results = Array.isArray(event.results) ? event.results : [];
-        state.suggestions = Array.isArray(event.suggestions) ? event.suggestions : [];
-        state.selectedResultIndex = indexSearchVisibleResults(state).length > 0 ? 0 : -1;
-        renderProjectIndexSearch(event.id);
-      }
-
-      function handleProjectIndexSearchError(event) {
-        const state = indexSearchStateMap.get(event.id);
-        if (!state || event.request_id !== state.inFlightRequestId) {
-          return;
-        }
-        state.searching = false;
-        state.inFlightRequestId = 0;
-        state.inFlightSignature = "";
-        state.error = event.message || "Project index search failed.";
-        renderProjectIndexSearch(event.id);
-      }
-
-      function renderProjectIndexWindows() {
-        for (const windowId of indexSearchStateMap.keys()) {
-          renderProjectIndexSearch(windowId);
-        }
-      }
-
-      function moveIndexResultSelection(windowId, delta) {
-        const state = ensureIndexSearchState(windowId);
-        const visibleResults = indexSearchVisibleResults(state);
-        if (visibleResults.length === 0) return;
-        const current = Math.max(0, state.selectedResultIndex);
-        state.selectedResultIndex = Math.max(0, Math.min(visibleResults.length - 1, current + delta));
-        renderProjectIndexSearch(windowId);
-        document
-          .querySelector(`[data-id='${CSS.escape(windowId)}'] [data-result-index='${state.selectedResultIndex}']`)
-          ?.focus();
-      }
-
-      function openIndexResultTarget(result) {
-        const target = result?.target || {};
-        switch (target.kind) {
-          case "issue":
-            openKnowledgeIndexResultTarget("issue", target);
-            return;
-          case "spec":
-            openKnowledgeIndexResultTarget("spec", target);
-            return;
-          case "board":
-            focusOrSpawnPreset("board");
-            return;
-          case "file":
-            focusOrSpawnPreset("file_tree");
-            return;
-          case "memory":
-          case "discussion":
-            focusOrSpawnPreset("index");
-            return;
-          default:
-            return;
-        }
-      }
-
-      function indexResultTargetNumber(target) {
-        const rawNumber = target?.number ?? target?.spec_id ?? target?.id;
-        const number = Number(rawNumber);
-        if (!Number.isInteger(number) || number <= 0) {
-          return null;
-        }
-        return number;
-      }
-
-      function openKnowledgeIndexResultTarget(preset, target) {
-        const knowledgeKind = knowledgeKindForPreset(preset);
-        const number = indexResultTargetNumber(target);
-        const windowId = focusOrSpawnPreset(preset);
-        if (!knowledgeKind || number === null) {
-          return;
-        }
-        if (windowId) {
-          requestKnowledgeDetail(windowId, knowledgeKind, number);
-          renderKnowledgeBridge(windowId);
-          return;
-        }
-        pendingIndexOpenTargetsByPreset.set(preset, { knowledgeKind, number });
       }
 
       function send(message) {
@@ -3859,460 +3352,14 @@
       }
 
       // ---- Provider usage & rate limits (SPEC-2970) ----
-      let latestProviderUsage = { accounts: [], sessions: [], consumption: [] };
-
-      const USAGE_PROVIDER_NAME = { codex: "Codex", claude_code: "Claude Code" };
-      const USAGE_WINDOW_LABEL = {
-        five_hour: "5-hour",
-        weekly: "Weekly",
-        opus_weekly: "Opus weekly",
-        sonnet_weekly: "Sonnet weekly",
-        code_review_weekly: "Code review weekly",
-      };
-
-      function usageStateReason(state) {
-        if (!state) return "";
-        switch (state.kind) {
-          case "disabled":
-            return "Enable in Settings";
-          case "no_data":
-            return "No data yet";
-          case "unavailable":
-            return state.reason ? `Unavailable — ${state.reason}` : "Unavailable";
-          case "stale":
-            return `stale ${Math.round((state.age_secs || 0) / 60)}m`;
-          default:
-            return "";
-        }
-      }
-
-      function usageFmtResetAt(iso) {
-        if (!iso) return "";
-        const d = new Date(iso);
-        if (Number.isNaN(d.getTime())) return "";
-        return d.toLocaleString();
-      }
-
-      function usageFmtTokens(n) {
-        if (n == null) return "—";
-        if (n >= 1000000) return `${(n / 1000000).toFixed(1)}M`;
-        if (n >= 1000) return `${Math.round(n / 1000)}k`;
-        return String(n);
-      }
-
-      function applyProviderUsageUi(snapshot) {
-        latestProviderUsage = snapshot || { accounts: [], sessions: [], consumption: [] };
-        try {
-          window.__operatorShell?.applyProviderUsage?.(latestProviderUsage);
-        } catch (e) {
-          console.warn("usage pill update failed", e);
-        }
-        try {
-          refreshUsageHoverIfOpen();
-        } catch {
-          /* no-op */
-        }
-        // Re-render regardless of session count: when a snapshot drops back to
-        // sessions:[] (agent stopped, rollout/transcript unreadable, settings
-        // change) the Work surface must clear its stale token/context instead
-        // of keeping the previous poll's values. SPEC-2359 Phase W-12 Slice 3
-        // (FR-351): the sidebar Active Works overview is gone, so usage now
-        // refreshes through the Workspace Overview (Kanban) Work surface.
-        try {
-          workspaceOverviewSurface.renderWindows();
-        } catch {
-          /* no-op */
-        }
-      }
-
-      function usageForSession(sessionId) {
-        return (
-          (latestProviderUsage.sessions || []).find(
-            (s) => s.session_id === sessionId,
-          ) || null
-        );
-      }
-
-      function buildUsageBar(percent) {
-        const wrap = document.createElement("div");
-        wrap.className = "op-usage-bar";
-        const fill = document.createElement("div");
-        fill.className = "op-usage-bar__fill";
-        const pct = Math.max(0, Math.min(100, Math.round(percent)));
-        fill.style.width = `${pct}%`;
-        if (pct >= 90) fill.dataset.level = "high";
-        else if (pct >= 70) fill.dataset.level = "mid";
-        wrap.appendChild(fill);
-        return wrap;
-      }
-
-      function renderUsageAccountRow(account) {
-        const row = document.createElement("div");
-        row.className = "op-usage-account";
-        row.dataset.provider = account.provider;
-        const head = document.createElement("div");
-        head.className = "op-usage-account__head";
-        const name = document.createElement("span");
-        name.className = "op-usage-account__name";
-        name.textContent = USAGE_PROVIDER_NAME[account.provider] || account.provider;
-        head.appendChild(name);
-        if (account.plan) {
-          const plan = document.createElement("span");
-          plan.className = "op-usage-account__plan";
-          plan.textContent = account.plan;
-          head.appendChild(plan);
-        }
-        const reason = usageStateReason(account.state);
-        if (reason) {
-          const isDisabled = (account.state && account.state.kind) === "disabled";
-          const r = document.createElement(isDisabled ? "button" : "span");
-          r.className = "op-usage-account__reason";
-          r.textContent = reason;
-          if (isDisabled) {
-            r.type = "button";
-            r.classList.add("op-usage-account__reason--action");
-            r.addEventListener("click", (e) => {
-              e.stopPropagation();
-              if (typeof window.__gwtHideUsageHover === "function") {
-                window.__gwtHideUsageHover();
-              }
-              document.dispatchEvent(
-                new CustomEvent("settings:open", { detail: { target: "usage" } }),
-              );
-            });
-          }
-          head.appendChild(r);
-        }
-        row.appendChild(head);
-        for (const w of account.windows || []) {
-          const line = document.createElement("div");
-          line.className = "op-usage-window";
-          const label = document.createElement("span");
-          label.className = "op-usage-window__label";
-          label.textContent = USAGE_WINDOW_LABEL[w.kind] || w.kind;
-          const pct = document.createElement("span");
-          pct.className = "op-usage-window__pct";
-          pct.textContent = `${Math.round(w.used_percent)}%`;
-          line.appendChild(label);
-          line.appendChild(buildUsageBar(w.used_percent));
-          line.appendChild(pct);
-          if (w.resets_at) {
-            const reset = document.createElement("span");
-            reset.className = "op-usage-window__reset";
-            reset.textContent = `↻ ${usageFmtResetAt(w.resets_at)}`;
-            line.appendChild(reset);
-          }
-          row.appendChild(line);
-        }
-        return row;
-      }
-
-      function consumptionTotal(b) {
-        if (!b) return 0;
-        return (b.input || 0) + (b.output || 0) + (b.cached || 0);
-      }
-
-      function fmtConsumptionBreakdown(b) {
-        if (!b) return "—";
-        return `in ${usageFmtTokens(b.input || 0)} · out ${usageFmtTokens(
-          b.output || 0,
-        )} · cached ${usageFmtTokens(b.cached || 0)}`;
-      }
-
-      function renderConsumptionChart(days) {
-        const chart = document.createElement("div");
-        chart.className = "op-usage-chart";
-        const totals = days.map((d) => consumptionTotal(d.breakdown));
-        const max = Math.max(1, ...totals);
-        days.forEach((d, i) => {
-          const col = document.createElement("div");
-          col.className = "op-usage-chart__col";
-          if (i === days.length - 1) col.dataset.today = "true";
-          const bar = document.createElement("div");
-          bar.className = "op-usage-chart__bar";
-          const total = totals[i];
-          bar.style.height = `${Math.max(2, Math.round((total / max) * 100))}%`;
-          bar.title = `${d.date}: ${usageFmtTokens(total)} tokens`;
-          col.appendChild(bar);
-          chart.appendChild(col);
-        });
-        return chart;
-      }
-
-      function usageFmtResetShort(iso) {
-        if (!iso) return "";
-        const d = new Date(iso);
-        if (Number.isNaN(d.getTime())) return "";
-        return d.toLocaleString(undefined, {
-          month: "numeric",
-          day: "numeric",
-          hour: "2-digit",
-          minute: "2-digit",
-        });
-      }
-
-      function usageConsumptionFor(provider) {
-        return (
-          (latestProviderUsage.consumption || []).find((c) => c.provider === provider) || null
-        );
-      }
-
-      // One rate-limit window as an aligned row: label · bar · % · reset.
-      function buildUsageWindowRow(w) {
-        const row = document.createElement("div");
-        row.className = "op-usage-win";
-        const lbl = document.createElement("span");
-        lbl.className = "op-usage-win__lbl";
-        lbl.textContent = USAGE_WINDOW_LABEL[w.kind] || w.kind;
-        const bar = buildUsageBar(w.used_percent);
-        bar.classList.add("op-usage-win__bar");
-        const pct = document.createElement("span");
-        pct.className = "op-usage-win__pct";
-        pct.textContent = `${Math.round(w.used_percent)}%`;
-        const reset = document.createElement("span");
-        reset.className = "op-usage-win__reset";
-        reset.textContent = w.resets_at ? `↻ ${usageFmtResetShort(w.resets_at)}` : "";
-        row.appendChild(lbl);
-        row.appendChild(bar);
-        row.appendChild(pct);
-        row.appendChild(reset);
-        return row;
-      }
-
-      // Consumption as an aligned 4-column grid (period × in/out/cached).
-      function buildUsageConsumptionGrid(pc) {
-        const grid = document.createElement("div");
-        grid.className = "op-usage-cgrid";
-        const t = pc.today || {};
-        const w = pc.this_week || {};
-        const cells = [
-          ["hdr", "tokens"],
-          ["colh", "in"],
-          ["colh", "out"],
-          ["colh", "cached"],
-          ["rowh", "Today"],
-          ["num", usageFmtTokens(t.input || 0)],
-          ["num", usageFmtTokens(t.output || 0)],
-          ["num", usageFmtTokens(t.cached || 0)],
-          ["rowh", "Week"],
-          ["num", usageFmtTokens(w.input || 0)],
-          ["num", usageFmtTokens(w.output || 0)],
-          ["num", usageFmtTokens(w.cached || 0)],
-        ];
-        for (const [kind, text] of cells) {
-          const cell = document.createElement("span");
-          cell.className = `op-usage-cgrid__${kind}`;
-          cell.textContent = text;
-          grid.appendChild(cell);
-        }
-        return grid;
-      }
-
-      // A provider card: header (icon · name · plan) + rate-limit windows (or a
-      // degraded reason) + consumption grid + 7-day sparkline. Grouping all of
-      // one provider's data together is the key readability win.
-      function buildUsageProviderCard(account) {
-        const card = document.createElement("div");
-        card.className = "op-usage-card";
-        card.dataset.provider = account.provider;
-
-        const head = document.createElement("div");
-        head.className = "op-usage-card__head";
-        const icon = document.createElement("span");
-        icon.className = "op-usage-card__icon";
-        icon.textContent = account.provider === "claude_code" ? "◇" : "⬡";
-        const name = document.createElement("span");
-        name.className = "op-usage-card__name";
-        name.textContent = USAGE_PROVIDER_NAME[account.provider] || account.provider;
-        head.appendChild(icon);
-        head.appendChild(name);
-        if (account.plan) {
-          const plan = document.createElement("span");
-          plan.className = "op-usage-card__plan";
-          plan.textContent = account.plan;
-          head.appendChild(plan);
-        }
-        card.appendChild(head);
-
-        const windows = account.windows || [];
-        if (windows.length) {
-          const wins = document.createElement("div");
-          wins.className = "op-usage-wins";
-          for (const w of windows) wins.appendChild(buildUsageWindowRow(w));
-          card.appendChild(wins);
-        } else {
-          const reason = usageStateReason(account.state);
-          if (reason) {
-            const isDisabled = (account.state && account.state.kind) === "disabled";
-            const r = document.createElement(isDisabled ? "button" : "div");
-            r.className = "op-usage-card__reason";
-            r.textContent = reason;
-            if (isDisabled) {
-              r.type = "button";
-              r.classList.add("op-usage-card__reason--action");
-              r.addEventListener("click", (e) => {
-                e.stopPropagation();
-                if (typeof window.__gwtHideUsageHover === "function") {
-                  window.__gwtHideUsageHover();
-                }
-                document.dispatchEvent(
-                  new CustomEvent("settings:open", { detail: { target: "usage" } }),
-                );
-              });
-            }
-            card.appendChild(r);
-          }
-        }
-
-        const pc = usageConsumptionFor(account.provider);
-        if (pc) {
-          const cwrap = document.createElement("div");
-          cwrap.className = "op-usage-card__cons";
-          cwrap.appendChild(buildUsageConsumptionGrid(pc));
-          if (Array.isArray(pc.days) && pc.days.length) {
-            cwrap.appendChild(renderConsumptionChart(pc.days));
-          }
-          card.appendChild(cwrap);
-        }
-        return card;
-      }
-
-      // SPEC-2970 — the full usage detail as provider cards, appended to a
-      // container. The hover popover is the single surface for all usage info
-      // (the click-open modal was removed per UX feedback). The per-session
-      // list was removed per UX feedback — it grew to hundreds of rows and
-      // overwhelmed the popover; per-session token usage still drives each
-      // agent's inline footer (see usageForSession).
-      function buildUsageFullSections(container) {
-        for (const account of latestProviderUsage.accounts || []) {
-          container.appendChild(buildUsageProviderCard(account));
-        }
-      }
-
-      // ---- Consolidated usage hover popover (SPEC-2970 UX) ----
-      // Hovering the status-bar USAGE cell shows EVERYTHING at once (both
-      // providers' windows with bars + full consumption with charts +
-      // sessions). The click-open modal was removed per UX feedback — the hover
-      // popover is the single surface. Move the cursor into it to scroll/read.
-      let usageHoverEl = null;
-      let usageHoverHideTimer = null;
-      let usageHoverAnchor = null;
-
-      function buildUsageHoverBody() {
-        const wrap = document.createElement("div");
-        wrap.className = "op-usage-hover__body";
-        const head = document.createElement("div");
-        head.className = "op-usage-hover__head";
-        head.textContent = "Usage & Limits";
-        wrap.appendChild(head);
-        buildUsageFullSections(wrap);
-        return wrap;
-      }
-
-      function positionUsageHover() {
-        if (!usageHoverEl || !usageHoverAnchor) return;
-        const r = usageHoverAnchor.getBoundingClientRect();
-        const w = usageHoverEl.offsetWidth;
-        const left = Math.max(8, Math.min(r.left, window.innerWidth - w - 8));
-        usageHoverEl.style.left = `${left}px`;
-        usageHoverEl.style.bottom = `${Math.max(8, window.innerHeight - r.top + 6)}px`;
-      }
-
-      function cancelUsageHoverHide() {
-        if (usageHoverHideTimer) {
-          clearTimeout(usageHoverHideTimer);
-          usageHoverHideTimer = null;
-        }
-      }
-
-      function refreshUsageHoverIfOpen() {
-        if (!usageHoverEl || usageHoverEl.hidden) return;
-        while (usageHoverEl.firstChild) usageHoverEl.removeChild(usageHoverEl.firstChild);
-        usageHoverEl.appendChild(buildUsageHoverBody());
-        positionUsageHover();
-      }
-
-      window.__gwtShowUsageHover = (anchor) => {
-        cancelUsageHoverHide();
-        usageHoverAnchor = anchor || usageHoverAnchor;
-        if (!usageHoverEl) {
-          usageHoverEl = document.createElement("div");
-          usageHoverEl.className = "op-usage-hover";
-          usageHoverEl.addEventListener("mouseenter", cancelUsageHoverHide);
-          usageHoverEl.addEventListener("mouseleave", () => window.__gwtHideUsageHover());
-          document.body.appendChild(usageHoverEl);
-        }
-        while (usageHoverEl.firstChild) usageHoverEl.removeChild(usageHoverEl.firstChild);
-        usageHoverEl.appendChild(buildUsageHoverBody());
-        usageHoverEl.hidden = false;
-        usageHoverEl.style.visibility = "hidden";
-        requestAnimationFrame(() => {
-          positionUsageHover();
-          if (usageHoverEl) usageHoverEl.style.visibility = "visible";
-        });
-      };
-
-      window.__gwtHideUsageHover = () => {
-        cancelUsageHoverHide();
-        usageHoverHideTimer = setTimeout(() => {
-          if (usageHoverEl) usageHoverEl.hidden = true;
-          usageHoverHideTimer = null;
-        }, 180);
-      };
-
-      // SPEC-2970 FR-009/FR-013 — Settings "Usage & Limits" panel: Claude
-      // account usage is opt-in (Keychain + network); Codex is local + auto.
-      function renderUsagePanel(panel) {
-        while (panel.firstChild) panel.removeChild(panel.firstChild);
-        const section = document.createElement("div");
-        section.className = "settings-section";
-
-        const heading = document.createElement("h3");
-        heading.textContent = "Provider Usage & Limits";
-        section.appendChild(heading);
-
-        const codexNote = document.createElement("p");
-        codexNote.className = "settings-hint";
-        codexNote.textContent =
-          "Codex usage is read from local session files automatically.";
-        section.appendChild(codexNote);
-
-        const label = document.createElement("label");
-        label.className = "settings-toggle";
-        const checkbox = document.createElement("input");
-        checkbox.type = "checkbox";
-        const claudeAccount = (latestProviderUsage.accounts || []).find(
-          (a) => a.provider === "claude_code",
-        );
-        checkbox.checked = !!(
-          claudeAccount &&
-          claudeAccount.state &&
-          claudeAccount.state.kind !== "disabled"
-        );
-        checkbox.addEventListener("change", () => {
-          try {
-            send({
-              kind: "set_claude_account_usage_enabled",
-              enabled: checkbox.checked,
-            });
-          } catch {
-            /* no-op */
-          }
-        });
-        const span = document.createElement("span");
-        span.textContent = "Show Claude Code account usage (5-hour / weekly)";
-        label.appendChild(checkbox);
-        label.appendChild(span);
-        section.appendChild(label);
-
-        const consent = document.createElement("p");
-        consent.className = "settings-hint";
-        consent.textContent =
-          "Off by default (opt-in). When enabled, Claude account usage reads your OAuth token from the Keychain / credentials file and requests usage from the Anthropic API (polled at most once every 3 minutes). While disabled, no Keychain read or network request happens. Per-session token usage is read locally and is not affected by this setting.";
-        section.appendChild(consent);
-
-        panel.appendChild(section);
-      }
+      // SPEC-3064 Phase 3 (E1): the usage snapshot state, formatters, hover
+      // popover, and Settings usage panel moved to
+      // /provider-usage-surface.js. app.js keeps the receive() and Settings
+      // call sites and wires them through this factory.
+      const { applyProviderUsageUi, renderUsagePanel } = createProviderUsageSurface({
+        send,
+        renderWorkspaceWindows: () => workspaceOverviewSurface.renderWindows(),
+      });
 
       function activeWorkFocusableAgents(work) {
         const agents = Array.isArray(work?.agents) ? work.agents : [];
@@ -4744,813 +3791,31 @@
         return /Chrome\//.test(ua);
       }
 
-      const SUPPORTED_IMAGE_PASTE_MIME_TYPES = new Set([
-        "image/png",
-        "image/jpeg",
-        "image/webp",
-      ]);
-      const MAX_FILE_DROP_COUNT = 16;
-      const FILE_DROP_AGENT_TARGET_MESSAGE = "Drop files on a running Agent window.";
-      const FILE_DROP_COUNT_MESSAGE = `Drop up to ${MAX_FILE_DROP_COUNT} files at once.`;
-      const FILE_DROP_UPLOAD_FAILURE_MESSAGE = "Could not upload dropped file.";
-      const IMAGE_PASTE_UPLOAD_FAILURE_MESSAGE = "Could not upload pasted image.";
-
-      function findClipboardImagePasteItem(items) {
-        for (const item of Array.from(items || [])) {
-          if (
-            item?.kind === "file" &&
-            SUPPORTED_IMAGE_PASTE_MIME_TYPES.has(item.type)
-          ) {
-            return item;
-          }
-        }
-        return null;
-      }
-
-      function defaultImagePasteFilename(mimeType) {
-        switch (mimeType) {
-          case "image/jpeg":
-            return "clipboard-image.jpg";
-          case "image/webp":
-            return "clipboard-image.webp";
-          case "image/png":
-          default:
-            return "clipboard-image.png";
-        }
-      }
-
-      function uploadFileFromImageBlob(blob, mimeType, filename) {
-        const type = mimeType || blob?.type || "";
-        const name = filename || blob?.name || defaultImagePasteFilename(type);
-        if (typeof File === "function" && (!(blob instanceof File) || !blob.name)) {
-          return new File([blob], name, { type });
-        }
-        return blob;
-      }
-
-      function dataTransferHasFiles(dataTransfer) {
-        const types = Array.from(dataTransfer?.types || []);
-        return types.includes("Files") || Boolean(dataTransfer?.files?.length);
-      }
-
-      function droppedFilesWithinCountLimit(files) {
-        return files.length <= MAX_FILE_DROP_COUNT;
-      }
-
-      function droppedFilesValidationFailure(files) {
-        if (!droppedFilesWithinCountLimit(files)) {
-          return FILE_DROP_COUNT_MESSAGE;
-        }
-        return null;
-      }
-
-      function showFileDropAlert(message) {
-        if (typeof window.alert === "function") {
-          window.alert(message);
-        }
-      }
-
-      function totalFileBytes(files) {
-        return files.reduce((total, file) => total + (file.size || 0), 0);
-      }
-
-      function displayAttachmentBasename(filename) {
-        const value = String(filename || "").trim();
-        const parts = value.split(/[\\/]+/).filter(Boolean);
-        return parts.at(-1) || "file";
-      }
-
-      function attachmentFileCountLabel(count) {
-        return `${count} ${count === 1 ? "file" : "files"}`;
-      }
-
-      function createAttachmentOperationId() {
-        const random =
-          typeof globalThis.crypto?.randomUUID === "function"
-            ? globalThis.crypto.randomUUID()
-            : `${Date.now().toString(36)}-${Math.random().toString(36).slice(2)}`;
-        return `attachment-${random}`;
-      }
-
-      const attachmentProgressControllers = new Map();
-
-      function ensureAttachmentProgressSurface(windowId) {
-        const host = workspaceWindowElement(windowId) || document.body;
-        let surface = host.querySelector(".attachment-progress");
-        if (surface) {
-          return surface;
-        }
-        surface = document.createElement("div");
-        surface.className = "attachment-progress";
-        surface.hidden = true;
-        surface.setAttribute("role", "status");
-        surface.setAttribute("aria-live", "polite");
-        surface.innerHTML = `
-          <div class="attachment-progress__row">
-            <span class="attachment-progress__label"></span>
-          </div>
-          <div class="attachment-progress__track" role="progressbar" aria-valuemin="0" aria-valuemax="100">
-            <div class="attachment-progress__bar"></div>
-          </div>
-        `;
-        host.appendChild(surface);
-        return surface;
-      }
-
-      function attachmentPhaseLabel(phase, fallback = "") {
-        switch (phase) {
-          case "queued":
-            return "Queued";
-          case "staging":
-            return "Staging";
-          case "injecting":
-            return "Injecting";
-          case "attached":
-            return "Attached";
-          case "failed":
-            return fallback || "Could not attach";
-          default:
-            return fallback || "Uploading";
-        }
-      }
-
-      function createAttachmentProgressController(windowId, files, operationId = createAttachmentOperationId()) {
-        const surface = ensureAttachmentProgressSurface(windowId);
-        const label = surface.querySelector(".attachment-progress__label");
-        const track = surface.querySelector(".attachment-progress__track");
-        const bar = surface.querySelector(".attachment-progress__bar");
-        const abortController = new AbortController();
-        const fileCount = files.length;
-        const state = {
-          phase: "Uploading",
-          filename: fileCount === 1 ? displayAttachmentBasename(files[0]?.name) : "",
-          totalBytes: totalFileBytes(files),
-          loadedBytes: 0,
-          failed: false,
-          done: false,
-          visible: true,
-        };
-
-        function percent() {
-          if (state.totalBytes <= 0) {
-            return state.done ? 100 : 0;
-          }
-          return Math.max(
-            0,
-            Math.min(100, Math.round((state.loadedBytes / state.totalBytes) * 100)),
-          );
-        }
-
-        function render() {
-          if (!state.visible) {
-            return;
-          }
-          const value = percent();
-          const filename = state.filename ? ` · ${state.filename}` : "";
-          const suffix = state.totalBytes > 0 ? ` · ${value}%` : "";
-          surface.hidden = false;
-          surface.dataset.state = state.failed ? "error" : state.done ? "done" : "active";
-          label.textContent = `${state.phase} ${attachmentFileCountLabel(fileCount)}${filename}${suffix}`;
-          track.setAttribute("aria-valuenow", String(value));
-          bar.style.width = `${value}%`;
-        }
-
-        function showNow() {
-          state.visible = true;
-          render();
-        }
-
-        render();
-
-        function setPhase(phase) {
-          state.phase = phase;
-          render();
-        }
-
-        function setUploadProgress(loadedBytes, totalBytes = null) {
-          if (Number.isFinite(totalBytes) && totalBytes >= 0) {
-            state.totalBytes = totalBytes;
-          }
-          state.loadedBytes = Math.max(0, loadedBytes || 0);
-          render();
-        }
-
-        function succeed() {
-          state.done = true;
-          state.phase = "Attached";
-          state.loadedBytes = state.totalBytes;
-          if (state.visible) {
-            render();
-            setTimeout(() => {
-              if (surface.dataset.state === "done") {
-                surface.hidden = true;
-                attachmentProgressControllers.delete(operationId);
-              }
-            }, 700);
-          }
-        }
-
-        function fail(message) {
-          state.failed = true;
-          state.phase = message;
-          showNow();
-        }
-
-        function applyBackendProgress(event) {
-          if (event.filename) {
-            state.filename = displayAttachmentBasename(event.filename);
-          }
-          if (Number.isFinite(event.bytes_total)) {
-            state.totalBytes = event.bytes_total;
-          }
-          if (Number.isFinite(event.bytes_done)) {
-            state.loadedBytes = event.bytes_done;
-          }
-          if (event.phase === "failed") {
-            fail(event.message || "Could not attach");
-            return;
-          }
-          if (event.phase === "attached") {
-            succeed();
-            return;
-          }
-          setPhase(attachmentPhaseLabel(event.phase));
-        }
-
-        const controller = {
-          operationId,
-          signal: abortController.signal,
-          setPhase,
-          setUploadProgress,
-          succeed,
-          fail,
-          applyBackendProgress,
-        };
-        attachmentProgressControllers.set(operationId, controller);
-        return controller;
-      }
-
-      function handleAttachmentProgress(event) {
-        const operationId = event?.operation_id || "";
-        if (!operationId) {
-          return;
-        }
-        let controller = attachmentProgressControllers.get(operationId);
-        if (!controller) {
-          controller = createAttachmentProgressController(
-            event.id,
-            [
-              {
-                name: event.filename || "file",
-                size: event.bytes_total || 0,
-              },
-            ],
-            operationId,
-          );
-        }
-        controller.applyBackendProgress(event);
-      }
-
-      function attachmentFilesFromNativePaths(paths) {
-        return paths.map((path) => ({
-          name: displayAttachmentBasename(path),
-          size: 0,
-        }));
-      }
-
-      let attachmentUploadTokenPromise = null;
-
-      async function attachmentUploadToken() {
-        if (!attachmentUploadTokenPromise) {
-          attachmentUploadTokenPromise = fetch("/internal/attachment-upload-token", {
-            credentials: "same-origin",
-          }).then(async (response) => {
-            if (!response.ok) {
-              throw new Error(`attachment upload token failed: ${response.status}`);
-            }
-            const payload = await response.json();
-            if (!payload?.token) {
-              throw new Error("attachment upload token missing");
-            }
-            return payload.token;
-          });
-        }
-        return attachmentUploadTokenPromise;
-      }
-
-      function uploadAttachmentFile(file, { onProgress, signal } = {}) {
-        if (typeof window.__gwtAttachmentUploader === "function") {
-          return window.__gwtAttachmentUploader({ file, onProgress, signal });
-        }
-        return new Promise((resolve, reject) => {
-          void attachmentUploadToken()
-            .then((token) => {
-              if (signal?.aborted) {
-                reject(new Error("upload aborted"));
-                return;
-              }
-              const params = new URLSearchParams({
-                filename: file.name || "file",
-                size: String(file.size || 0),
-              });
-              if (file.type) {
-                params.set("mime_type", file.type);
-              }
-              const xhr = new XMLHttpRequest();
-              xhr.open("POST", `/internal/attachments/upload?${params.toString()}`);
-              xhr.setRequestHeader("x-gwt-upload-token", token);
-              xhr.responseType = "json";
-              xhr.upload.onprogress = (event) => {
-                onProgress?.({
-                  loaded: event.loaded || 0,
-                  total: event.lengthComputable ? event.total : file.size || 0,
-                });
-              };
-              xhr.onload = () => {
-                if (xhr.status < 200 || xhr.status >= 300) {
-                  reject(new Error(`upload failed: ${xhr.status}`));
-                  return;
-                }
-                resolve(xhr.response || JSON.parse(xhr.responseText || "{}"));
-              };
-              xhr.onerror = () => reject(new Error("upload failed"));
-              xhr.onabort = () => reject(new Error("upload aborted"));
-              signal?.addEventListener("abort", () => xhr.abort(), { once: true });
-              xhr.send(file);
-            })
-            .catch(reject);
-        });
-      }
-
-      async function uploadFilesAsAttachments(files, progress) {
-        const totalBytes = totalFileBytes(files);
-        let completedBytes = 0;
-        const attachments = [];
-        for (const file of files) {
-          const uploaded = await uploadAttachmentFile(file, {
-            signal: progress.signal,
-            onProgress: ({ loaded }) => {
-              progress.setUploadProgress(completedBytes + (loaded || 0), totalBytes);
-            },
-          });
-          completedBytes += file.size || uploaded?.size || 0;
-          progress.setUploadProgress(
-            totalBytes > 0 ? Math.min(completedBytes, totalBytes) : 0,
-            totalBytes,
-          );
-          attachments.push({
-            source: "uploaded",
-            upload_id: uploaded.upload_id,
-            filename: uploaded.filename || file.name || "file",
-            mime_type: uploaded.mime_type ?? (file.type || null),
-            size: uploaded.size ?? file.size ?? 0,
-          });
-        }
-        return attachments;
-      }
-
-      async function uploadPastedImage(windowId, blob, { mimeType, filename } = {}) {
-        const file = uploadFileFromImageBlob(blob, mimeType, filename);
-        const progress = createAttachmentProgressController(windowId, [file]);
-        try {
-          const uploaded = await uploadAttachmentFile(file, {
-            signal: progress.signal,
-            onProgress: ({ loaded, total }) => progress.setUploadProgress(loaded || 0, total),
-          });
-          progress.setPhase("Queued");
-          send({
-            kind: "paste_image_uploaded",
-            id: windowId,
-            operation_id: progress.operationId,
-            upload_id: uploaded.upload_id,
-            mime_type: uploaded.mime_type ?? file.type ?? mimeType ?? "",
-            filename: uploaded.filename || file.name || filename || null,
-            size: uploaded.size ?? file.size ?? 0,
-          });
-        } catch (_error) {
-          progress.fail(IMAGE_PASTE_UPLOAD_FAILURE_MESSAGE);
-          showFileDropAlert(IMAGE_PASTE_UPLOAD_FAILURE_MESSAGE);
-        }
-      }
-
-      async function readNavigatorClipboardItems() {
-        if (!navigator.clipboard?.read) {
-          return [];
-        }
-        return navigator.clipboard.read();
-      }
-
-      async function readNavigatorClipboardText() {
-        if (!navigator.clipboard?.readText) {
-          return "";
-        }
-        return navigator.clipboard.readText();
-      }
-
-      async function writeClipboardText(text, restoreFocus = null) {
-        if (!text) {
-          return false;
-        }
-        if (navigator.clipboard?.writeText) {
-          try {
-            await navigator.clipboard.writeText(text);
-            restoreFocus?.();
-            return true;
-          } catch (_error) {
-            // Fall back to a temporary textarea when the async clipboard API is unavailable.
-          }
-        }
-
-        const textarea = document.createElement("textarea");
-        textarea.value = text;
-        textarea.setAttribute("readonly", "");
-        textarea.style.position = "fixed";
-        textarea.style.top = "-1000px";
-        textarea.style.left = "-1000px";
-        textarea.style.opacity = "0";
-        document.body.appendChild(textarea);
-        textarea.focus();
-        textarea.select();
-
-        try {
-          return document.execCommand("copy");
-        } catch (_error) {
-          return false;
-        } finally {
-          textarea.remove();
-          restoreFocus?.();
-        }
-      }
-
-      async function copyTerminalSelection(windowId, { clearSelectionAfterCopy = false } = {}) {
-        const runtime = terminalMap.get(windowId);
-        if (!runtime || !runtime.terminal.hasSelection()) {
-          return false;
-        }
-        const selection = runtime.terminal.getSelection();
-        if (!selection) {
-          return false;
-        }
-        const copied = await writeClipboardText(selection, () => runtime.terminal.focus());
-        if (copied && clearSelectionAfterCopy) {
-          runtime.terminal.clearSelection();
-        }
-        return copied;
-      }
-
-      async function copyTerminalOverlayMessage(windowId) {
-        const element = windowMap.get(windowId);
-        const messageEl = element?.querySelector(".terminal-overlay .overlay-message");
-        if (!messageEl) {
-          return false;
-        }
-        return writeClipboardText(messageEl.textContent, () => {
-          terminalMap.get(windowId)?.terminal.focus();
-        });
-      }
-
-      function updateTerminalOverlayCopyState(overlay) {
-        const button = overlay?.querySelector(".overlay-copy-button");
-        const messageEl = overlay?.querySelector(".overlay-message");
-        if (!button || !messageEl) {
-          return;
-        }
-        const hasMessage = Boolean(messageEl.textContent);
-        button.hidden = !hasMessage;
-        button.disabled = !hasMessage;
-      }
-
-      function installTerminalCopyHandlers(windowId, terminalRoot, terminal) {
-        const copyState = {
-          mouseDown: false,
-          dragged: false,
-          startX: 0,
-          startY: 0,
-        };
-
-        function resetCopyState() {
-          copyState.mouseDown = false;
-          copyState.dragged = false;
-          copyState.startX = 0;
-          copyState.startY = 0;
-        }
-
-        const handleMouseDown = (event) => {
-          if (event.button !== 0) {
-            return;
-          }
-          copyState.mouseDown = true;
-          copyState.dragged = false;
-          copyState.startX = event.clientX;
-          copyState.startY = event.clientY;
-        };
-
-        const handleMouseMove = (event) => {
-          if (!copyState.mouseDown || copyState.dragged) {
-            return;
-          }
-          const movedX = Math.abs(event.clientX - copyState.startX);
-          const movedY = Math.abs(event.clientY - copyState.startY);
-          if (
-            movedX >= TERMINAL_SELECTION_DRAG_THRESHOLD ||
-            movedY >= TERMINAL_SELECTION_DRAG_THRESHOLD
-          ) {
-            copyState.dragged = true;
-          }
-        };
-
-        const handleMouseUp = (event) => {
-          if (!copyState.mouseDown) {
-            return;
-          }
-          const shouldCopy = event.button === 0 && copyState.dragged;
-          resetCopyState();
-          if (!shouldCopy) {
-            return;
-          }
-          requestAnimationFrame(() => {
-            if (!terminal.hasSelection()) {
-              return;
-            }
-            void copyTerminalSelection(windowId);
-          });
-        };
-
-        const handleWindowBlur = () => {
-          resetCopyState();
-        };
-
-        terminal.attachCustomKeyEventHandler((event) => {
-          const copyDecision = classifyTerminalCopyKeyEvent(event, {
-            hasSelection: terminal.hasSelection(),
-          });
-          if (!copyDecision.copy) {
-            return true;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-          if (!terminal.hasSelection()) {
-            return false;
-          }
-          void copyTerminalSelection(windowId, {
-            clearSelectionAfterCopy: copyDecision.clearSelectionAfterCopy,
-          });
-          return false;
-        });
-
-        terminalRoot.addEventListener("mousedown", handleMouseDown);
-        window.addEventListener("mousemove", handleMouseMove, true);
-        window.addEventListener("mouseup", handleMouseUp, true);
-        window.addEventListener("blur", handleWindowBlur);
-
-        return () => {
-          terminal.attachCustomKeyEventHandler(() => true);
-          terminalRoot.removeEventListener("mousedown", handleMouseDown);
-          window.removeEventListener("mousemove", handleMouseMove, true);
-          window.removeEventListener("mouseup", handleMouseUp, true);
-          window.removeEventListener("blur", handleWindowBlur);
-        };
-      }
-
-      function installTerminalImagePasteHandlers(windowId, terminalRoot, terminal) {
-        const handlePaste = (event) => {
-          const item = findClipboardImagePasteItem(event.clipboardData?.items);
-          if (!item) {
-            return;
-          }
-          const file = item.getAsFile?.();
-          if (!file) {
-            return;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-
-          void uploadPastedImage(windowId, file, {
-            mimeType: file.type || item.type,
-            filename: file.name || null,
-          }).finally(() => {
-            terminal.focus();
-          });
-        };
-
-        terminalRoot.addEventListener("paste", handlePaste, true);
-        return () => {
-          terminalRoot.removeEventListener("paste", handlePaste, true);
-        };
-      }
-
-      function installTerminalFileDropHandlers(windowId, terminalRoot, terminal) {
-        const handleDragOver = (event) => {
-          if (!dataTransferHasFiles(event.dataTransfer)) {
-            return;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-          event.dataTransfer.dropEffect = "copy";
-        };
-
-        const handleDrop = (event) => {
-          const files = Array.from(event.dataTransfer?.files || []);
-          if (files.length === 0) {
-            return;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-          if (!isAgentWindowPreset(workspaceWindowById(windowId)?.preset)) {
-            showFileDropAlert(FILE_DROP_AGENT_TARGET_MESSAGE);
-            terminal.focus();
-            return;
-          }
-          const failure = droppedFilesValidationFailure(files);
-          if (failure) {
-            showFileDropAlert(failure);
-            terminal.focus();
-            return;
-          }
-
-          void sendDroppedFileAttachments(windowId, files).finally(() => {
-            terminal.focus();
-          });
-        };
-
-        terminalRoot.addEventListener("dragover", handleDragOver, true);
-        terminalRoot.addEventListener("drop", handleDrop, true);
-        return () => {
-          terminalRoot.removeEventListener("dragover", handleDragOver, true);
-          terminalRoot.removeEventListener("drop", handleDrop, true);
-        };
-      }
-
-      function installTerminalContextMenuHandlers(windowId, terminalRoot, terminal) {
-        const controller = createTerminalContextMenuController({
-          document,
-          window,
-          terminalRoot,
-          readClipboardText: readNavigatorClipboardText,
-          readClipboardItems: readNavigatorClipboardItems,
-          supportedImageTypes: SUPPORTED_IMAGE_PASTE_MIME_TYPES,
-          pasteText: (text) => terminal.paste(text),
-          pasteImage: ({ blob, mimeType, filename }) =>
-            uploadPastedImage(windowId, blob, { mimeType, filename }),
-          focusTerminal: () => terminal.focus(),
-        });
-        return () => {
-          controller.dispose();
-        };
-      }
-
-      function installTerminalViewportRefreshHandlers(windowId, terminal) {
-        const viewportScrollDisposable = terminal.onScroll(() => {
-          scheduleTerminalViewportRefresh(windowId);
-        });
-
-        return () => {
-          viewportScrollDisposable.dispose();
-        };
-      }
-
-      function terminalWindowIdFromPoint(x, y) {
-        if (!Number.isFinite(x) || !Number.isFinite(y)) {
-          return null;
-        }
-        const target = document.elementFromPoint(x, y);
-        const terminalRoot = target?.closest?.(".terminal-root");
-        const windowElement = terminalRoot?.closest?.(".workspace-window");
-        const windowId = windowElement?.dataset?.id || null;
-        if (!windowId || !terminalMap.has(windowId)) {
-          return null;
-        }
-        return windowId;
-      }
-
-      function workspaceWindowIdFromDropEvent(event) {
-        const targetWindow = event.target?.closest?.(".workspace-window");
-        const targetWindowId = targetWindow?.dataset?.id || null;
-        if (targetWindowId) {
-          return targetWindowId;
-        }
-        const x = Number(event.clientX);
-        const y = Number(event.clientY);
-        if (!Number.isFinite(x) || !Number.isFinite(y)) {
-          return null;
-        }
-        const pointTarget = document.elementFromPoint(x, y);
-        const pointWindow = pointTarget?.closest?.(".workspace-window");
-        return pointWindow?.dataset?.id || null;
-      }
-
-      function agentWindowIdFromDropEvent(event) {
-        const windowId = workspaceWindowIdFromDropEvent(event);
-        if (!windowId || !terminalMap.has(windowId)) {
-          return null;
-        }
-        const windowData = workspaceWindowById(windowId);
-        if (!isAgentWindowPreset(windowData?.preset)) {
-          return null;
-        }
-        return windowId;
-      }
-
-      function eventTargetsTerminalRoot(event) {
-        return Boolean(event.target?.closest?.(".terminal-root"));
-      }
-
-      function focusTerminalForWindow(windowId) {
-        terminalMap.get(windowId)?.terminal.focus();
-      }
-
-      async function sendDroppedFileAttachments(windowId, files) {
-        const failure = droppedFilesValidationFailure(files);
-        if (failure) {
-          showFileDropAlert(failure);
-          focusTerminalForWindow(windowId);
-          return;
-        }
-
-        const progress = createAttachmentProgressController(windowId, files);
-        try {
-          const attachments = await uploadFilesAsAttachments(files, progress);
-          progress.setPhase("Queued");
-          send({
-            kind: "attach_files",
-            id: windowId,
-            operation_id: progress.operationId,
-            files: attachments,
-          });
-        } catch (_error) {
-          progress.fail(FILE_DROP_UPLOAD_FAILURE_MESSAGE);
-          showFileDropAlert(FILE_DROP_UPLOAD_FAILURE_MESSAGE);
-        } finally {
-          focusTerminalForWindow(windowId);
-        }
-      }
-
-      function installBrowserFileDropBridge() {
-        const handleDragOver = (event) => {
-          if (!dataTransferHasFiles(event.dataTransfer) || eventTargetsTerminalRoot(event)) {
-            return;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-          if (agentWindowIdFromDropEvent(event)) {
-            event.dataTransfer.dropEffect = "copy";
-          } else {
-            event.dataTransfer.dropEffect = "none";
-          }
-        };
-
-        const handleDrop = (event) => {
-          if (!dataTransferHasFiles(event.dataTransfer) || eventTargetsTerminalRoot(event)) {
-            return;
-          }
-          const files = Array.from(event.dataTransfer?.files || []);
-          if (files.length === 0) {
-            return;
-          }
-          event.preventDefault();
-          event.stopPropagation();
-          const windowId = agentWindowIdFromDropEvent(event);
-          if (!windowId) {
-            showFileDropAlert(FILE_DROP_AGENT_TARGET_MESSAGE);
-            return;
-          }
-          void sendDroppedFileAttachments(windowId, files);
-        };
-
-        window.addEventListener("dragover", handleDragOver, true);
-        window.addEventListener("drop", handleDrop, true);
-      }
-
-      function installNativeFileDropBridge() {
-        window.addEventListener("gwt:native-file-drop", (event) => {
-          const detail = event.detail || {};
-          const paths = Array.isArray(detail.paths)
-            ? detail.paths.filter((path) => typeof path === "string" && path.length > 0)
-            : [];
-          if (paths.length === 0) {
-            return;
-          }
-          const windowId = terminalWindowIdFromPoint(Number(detail.x), Number(detail.y));
-          if (!windowId) {
-            return;
-          }
-          const progress = createAttachmentProgressController(
-            windowId,
-            attachmentFilesFromNativePaths(paths),
-          );
-          progress.setPhase("Queued");
-          send({
-            kind: "attach_files",
-            id: windowId,
-            operation_id: progress.operationId,
-            files: paths.map((path) => ({
-              source: "native_path",
-              path,
-            })),
-          });
-          terminalMap.get(windowId)?.terminal.focus();
-        });
-      }
+      // SPEC-3064 Phase 3 (E2): terminal attachment, clipboard, and file-drop
+      // handling moved to /terminal-attachments.js. createTerminalRuntime,
+      // applyStatus, mountWindowBody, receive(), and the boot-time bridge
+      // installs keep their existing call sites wired through this factory.
+      const {
+        updateTerminalOverlayCopyState,
+        copyTerminalOverlayMessage,
+        installTerminalCopyHandlers,
+        installTerminalImagePasteHandlers,
+        installTerminalFileDropHandlers,
+        installTerminalContextMenuHandlers,
+        installTerminalViewportRefreshHandlers,
+        handleAttachmentProgress,
+        installBrowserFileDropBridge,
+        installNativeFileDropBridge,
+      } = createTerminalAttachments({
+        send,
+        terminalMap,
+        windowMap,
+        workspaceWindowById,
+        isAgentWindowPreset,
+        workspaceWindowElement,
+        scheduleTerminalViewportRefresh,
+        TERMINAL_SELECTION_DRAG_THRESHOLD,
+      });
 
       // SPEC-2356 — xterm content stays on the Dark Operator palette even when
       // surrounding window chrome follows the app's light theme. High-contrast
@@ -12037,160 +10302,7 @@
         }
 
         if (surface === "index") {
-          body.innerHTML = `
-            <div class="index-search-root">
-              <div class="index-window-header">
-                <div class="index-window-title">
-                  <div class="knowledge-heading">Index</div>
-                  <div class="index-window-subtitle">Search indexed project content</div>
-                </div>
-                <div class="workspace-toolbar-actions index-window-tabs">
-                  <button class="settings-tab active" type="button" role="tab" aria-selected="true" data-index-tab="search">Search</button>
-                  <button class="settings-tab" type="button" role="tab" aria-selected="false" data-index-tab="health">Health</button>
-                </div>
-              </div>
-              <section class="index-search-panel" data-index-panel="search">
-                <div class="index-search-toolbar">
-                  <form class="index-search-box" role="search">
-                    <input class="index-search-input" type="search" placeholder="Search by meaning, e.g. workspace lifecycle" aria-label="Search indexed content" />
-                    <button class="index-run-button" type="submit">Search</button>
-                  </form>
-                  <div class="index-filter-bar">
-                    <div class="index-match-mode-list" role="group" aria-label="Search mode">
-                      <button class="index-match-mode-button active" type="button" aria-pressed="true" data-match-mode="semantic">Semantic</button>
-                      <button class="index-match-mode-button" type="button" aria-pressed="false" data-match-mode="all_terms">All terms</button>
-                    </div>
-                    <div class="index-scope-list" role="group" aria-label="Search scopes"></div>
-                    <label class="index-worktree-field">
-                      <span>File worktree</span>
-                      <select class="index-worktree-select" aria-label="Files and Docs worktree"></select>
-                    </label>
-                  </div>
-                </div>
-                <div class="index-search-status"></div>
-                <div class="index-search-layout workspace-split">
-                  <div class="index-result-list workspace-scroll"></div>
-                  <div class="index-result-detail workspace-scroll"></div>
-                </div>
-              </section>
-              <section class="index-health-panel" data-index-panel="health" hidden>
-                <div class="index-health-toolbar">
-                  <div>
-                    <div class="index-health-title">Project index health</div>
-                    <div class="index-health-subtitle">Repair indexed sources without leaving this window.</div>
-                  </div>
-                  <button class="icon-button" data-action="refresh-index-status" aria-label="Refresh index status">↻</button>
-                </div>
-                <div class="index-health-table workspace-scroll"></div>
-              </section>
-            </div>
-          `;
-          body.addEventListener("mousedown", () => {
-            focusWindowLocally(windowData.id);
-            socketTransport.send({ kind: "focus_window", id: windowData.id });
-          });
-          const state = ensureIndexSearchState(windowData.id);
-          const input = body.querySelector(".index-search-input");
-          input.value = state.query;
-          input.addEventListener("input", () => {
-            state.query = input.value;
-            if (!state.query.trim()) {
-              clearProjectIndexSearchState(state);
-              renderProjectIndexSearch(windowData.id);
-              return;
-            }
-            markProjectIndexSearchPending(state);
-            renderProjectIndexSearch(windowData.id);
-            scheduleProjectIndexSearch(windowData.id);
-          });
-          body.querySelector(".index-search-box").addEventListener("submit", (event) => {
-            event.preventDefault();
-            if (state.searchTimer) {
-              clearTimeout(state.searchTimer);
-              state.searchTimer = 0;
-            }
-            sendProjectIndexSearch(windowData.id);
-          });
-          body.querySelector(".index-scope-list").addEventListener("click", (event) => {
-            const button = event.target.closest("[data-scope]");
-            if (!button) return;
-            const scope = button.dataset.scope;
-            if (state.selectedScopes.has(scope)) {
-              state.selectedScopes.delete(scope);
-            } else {
-              state.selectedScopes.add(scope);
-            }
-            if (state.query.trim()) {
-              markProjectIndexSearchPending(state);
-            }
-            renderProjectIndexSearch(windowData.id);
-            scheduleProjectIndexSearch(windowData.id);
-          });
-          body.querySelector(".index-match-mode-list").addEventListener("click", (event) => {
-            const button = event.target.closest("[data-match-mode]");
-            if (!button) return;
-            state.matchMode = button.dataset.matchMode || "semantic";
-            if (state.query.trim()) {
-              markProjectIndexSearchPending(state);
-            }
-            renderProjectIndexSearch(windowData.id);
-            scheduleProjectIndexSearch(windowData.id);
-          });
-          body.querySelector(".index-worktree-select").addEventListener("change", (event) => {
-            state.selectedWorktreeHash = event.target.value || "";
-            if (state.query.trim()) {
-              markProjectIndexSearchPending(state);
-              renderProjectIndexSearch(windowData.id);
-            }
-            scheduleProjectIndexSearch(windowData.id);
-          });
-          for (const tab of body.querySelectorAll("[data-index-tab]")) {
-            tab.addEventListener("click", () => {
-              state.activeTab = tab.dataset.indexTab || "search";
-              if (state.activeTab === "health") {
-                requestFullIndexStatusRefresh();
-              }
-              renderProjectIndexSearch(windowData.id);
-            });
-          }
-          body
-            .querySelector("[data-action='refresh-index-status']")
-            .addEventListener("click", (event) => {
-              event.stopPropagation();
-              requestFullIndexStatusRefresh();
-            });
-          body.querySelector(".index-result-list").addEventListener("click", (event) => {
-            const row = event.target.closest("[data-result-index]");
-            if (!row) return;
-            state.selectedResultIndex = Number(row.dataset.resultIndex || 0);
-            renderProjectIndexSearch(windowData.id);
-          });
-          body.querySelector(".index-result-list").addEventListener("dblclick", (event) => {
-            const row = event.target.closest("[data-result-index]");
-            if (!row) return;
-            state.selectedResultIndex = Number(row.dataset.resultIndex || 0);
-            const result = selectedIndexSearchItem(state)?.result;
-            openIndexResultTarget(result);
-          });
-          body.querySelector(".index-result-list").addEventListener("keydown", (event) => {
-            if (event.key === "ArrowDown") {
-              event.preventDefault();
-              moveIndexResultSelection(windowData.id, 1);
-            } else if (event.key === "ArrowUp") {
-              event.preventDefault();
-              moveIndexResultSelection(windowData.id, -1);
-            } else if (event.key === "Enter") {
-              event.preventDefault();
-              const result = selectedIndexSearchItem(state)?.result;
-              openIndexResultTarget(result);
-            }
-          });
-          body.querySelector(".index-result-detail").addEventListener("click", (event) => {
-            if (!event.target.closest("[data-action='open-index-result']")) return;
-            const result = selectedIndexSearchItem(state)?.result;
-            openIndexResultTarget(result);
-          });
-          renderProjectIndexSearch(windowData.id);
+          mountProjectIndexSurface(body, windowData);
           return;
         }
